Added sqlsrv support for the sqlsrv libaries
DLIB isn't supported on windows systems since php 5.3.
The official sqlsrv libaries and SQL server queries do require some
tweaks/configuration. These changes reflect that.
I Also added some expansion of the cryptic error messages SQL server
can spew out, to give some guidance to the beginning/I forgot about that
developer that's setting up a new system. These are the common pitfalls
that need to be checked, tried to make it work(saves on average 4 hours
of googling and trying). See it as an octoberCMS service to those who
struggle with SQL server :-)
If you choose to remove it, please consider adding it to some documentation
although a developer that is setting up a new system usually does not read
the documentation... he hits google immedeatly.
That is why I choose to expand the exception message.